+++ stdnet/orm/signals.py
@@ -1,5 +1,14 @@
 from stdnet.dispatch import Signal
 
+__all__ = ['class_prepared',
+           'pre_init',
+           'post_init',
+           'pre_save',
+           'post_save',
+           'pre_delete',
+           'post_delete']
+
+
 class_prepared = Signal(providing_args=["class"])
 
 pre_init = Signal(providing_args=["instance", "args", "kwargs"])